refactor: move out and export testing util
